Run 'god:restart' after 'deploy:restart'.
The 'deploy:restart' task is run at the end of a normal capistrano
deploy as well as at the end of a deploy:rollback. Therefore, it is a
better target for running god:restart than deploy. This will ensure that
god is restart both after a deploy and after a manual rollback.